Added EquipmentSet to models.
