+++ ambulance/models.py
@@ -162,7 +162,7 @@ class Ambulance(UpdatedByModel):
         loaded_values = self._loaded_values is not None
 
         # create equipment holder?
-        if created:
+        if created and self.equipmentholder is None:
             self.equipmentholder = EquipmentHolder.objects.create()
 
         # has location changed?
